requirements.txt
# ECG-Benchmark Requirements
# Python 3.8+

# Core Deep Learning
torch>=1.9.0
torchvision>=0.10.0
torchaudio>=0.9.0

# Data Processing & Analysis
numpy>=1.19.0
pandas>=1.1.0
scipy>=1.5.0
scikit-learn>=0.24.0

# ECG/Biophysical Signal Processing
wfdb>=3.4.0
mne>=0.22.0

# Time Series & Machine Learning
aeon>=0.10.0
hydra-core>=1.1.0
omegaconf>=2.1.0

# Visualization & Logging
matplotlib>=3.3.0
wandb>=0.12.0
seaborn>=0.11.0

# Utilities
PyYAML>=5.3.0
tqdm>=4.50.0
joblib>=1.0.0

# Development Tools (optional)
ipython>=7.0.0
jupyter>=1.0.0
pytest>=6.0.0
